// src/shared/types.ts
export type Id = string | number;
// ...
export type GroupStatus = 'Ativo' | 'Lotado' | 'Arquivado';

export interface Group {
  id: Id;
  name: string;
  description?: string;
  accessLink?: string;
  status: GroupStatus;
  memberGoal?: number;
}
// ...
export type CreateGroupData = Omit<Group, 'id'>;
export type UpdateGroupData = Partial<CreateGroupData>;